commited from goinfree again

+++ Request/Request.hpp
@@ -58,7 +58,7 @@ class Request {
         std::string fileName;
         // std::string bodySaver;
         struct stat pathStatus;
-        int contentLength; 
+        double contentLength; 
         int readbytes;
         int readBody;
         int chunked_flag;
